+++ V2/Variable.go
@@ -18,6 +18,6 @@ func (v Variable) isSolvable() bool {
 func (v Variable) print() {
 	fmt.Print(v.name)
 }
-func (v Variable) getSimplify() int {
+func (v Variable) getSimplify() float64 {
 	return SimplifyVariable
 }
